checking if a directory exists in post process

+++ server.js
@@ -23,12 +23,13 @@ app.get('/', (req, res) => {
 });
 
 app.post('/', async (req, res) => {
-  console.log(path.join(__dirname, './public/temp/'));
-  const files = fs.readdirSync(path.join(__dirname, './public/temp/'));
-  for (const file of files) {
-    await fs.unlink(path.join(__dirname, `./public/temp/${file}`), (err) => {
-      if (err) throw err;
-    });
+  if (fs.existsSync(path.join(__dirname, `./public/temp/`))) {
+    const files = fs.readdirSync(path.join(__dirname, './public/temp/'));
+    for (const file of files) {
+      await fs.unlink(path.join(__dirname, `./public/temp/${file}`), (err) => {
+        if (err) throw err;
+      });
+    }
   }
   data = [];
   data.push(req.body);
@@ -61,7 +62,6 @@ app.get('/download/dxf', async (req, res) => {
   res.download('./public/temp/' + fileDownload, (err) => {
     if (err) {
       console.log('Download error: ', err);
-      console.log(path.join(__dirname, './public/temp/'));
       res.redirect('/');
     }
   });
@@ -79,7 +79,6 @@ app.get('/download/svg', async (req, res) => {
   res.download('./public/temp/' + fileDownload, (err) => {
     if (err) {
       console.log('Download error: ', err);
-      console.log(path.join(__dirname, './public/temp/'));
 
       res.redirect('/');
     }
